Add requirements and move files to src
