+++ requirements.txt
@@ -1,0 +1,1 @@
+pygame==1.9.4
